fix: load handleViteDevServer adapter only in dev (#55)
